moved CSD info to csd_info.py
